fix: parse Safari versions correctly in `generator-networks-creator` (#347)
The `generator-networks-creator` package is parsing the browser versions from the `user-agent` string. 

While other browsers use the `chrome/130.0.0.0` suffix in the `user-agent` string, Safari uses the `Version/18.0 Safari/650.0.0.0` syntax - with `650.0.0.0` being the WebKit, not the Safari version.

This PR fixes this omission in `generator-networks-creator` and regenerates the generative models. There is also minor refactoring.