feat: add the `priority` header to compatible browser header sets (#367)
Injects a static `priority` header into compatible header sets (based on
the compatibility header from
[MDN](https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Headers/Priority)).

Generating this field based on the collected data isn't feasible, since
the Priority header is resource-type bound (is the resource a document,
image, stylesheet...?) rather than `user-agent`dependent. In this
regard, it's similar to the `sec-fetch-*` headers.

This PR sets the `priority` header to `u=0`, setting the highest
possible priority for the current resource.

Closes #351